# perf: avoiding use of Object.defineProperty in Chunk constructor